pyproject.toml
[tool.poetry]
name = "ssb-dapla-geoio"
packages = [{include = "dapla_geoio", from = "src"}]
version = "0.0.7"
description = "Dapla GeoIO"
authors = ["Bjørn Lie Rapp <bjorn.rapp@ssb.no>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/statisticsnorway/dapla-geoio"
repository = "https://github.com/statisticsnorway/dapla-geoio"
documentation = "https://statisticsnorway.github.io/dapla-geoio"
classifiers = ["Development Status :: 2 - Pre-Alpha"]

[tool.poetry.urls]
Changelog = "https://github.com/statisticsnorway/dapla-geoio/releases"

[tool.poetry.dependencies]
python = "^3.10"
geopandas = "^1.0.1"
pyogrio = "^0.10.0"
dapla-toolbelt = "^3.2.2"
strenum = "^0.4.15"
universal-pathlib = "^0.2.6"

[tool.poetry.group.dev.dependencies]
pygments = ">=2.10.0"
black = { extras = ["jupyter"], version = ">=23.1.0" }
coverage = { extras = ["toml"], version = ">=6.2" }
darglint = ">=1.8.1"
mypy = ">=0.930"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
ruff = ">=0.0.284"
typeguard = ">=2.13.3"
xdoctest = { extras = ["colors"], version = ">=0.15.10" }

[tool.poetry.group.test.dependencies]
pytest = ">=6.2.5"
pytest-mock = "^3.14.0"

[tool.poetry.group.type-stubs.dependencies]
types-shapely = "^2.0.0.20240618"
pandas-stubs = "^2.2.3.241126"
pyarrow-stubs = "^17.16"

[tool.poetry.group.docs.dependencies]
sphinx = ">=6.2.1"
sphinx-autobuild = ">=2021.3.14"
sphinx-autodoc-typehints = ">=1.24.0"
sphinx-click = ">=3.0.2"
furo = ">=2021.11.12"
myst-parser = { version = ">=0.16.1" }

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.poetry.scripts]
ssb-dapla-geoio = "dapla_geoio.__main__:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["dapla_geoio", "tests"]
relative_files = true

[tool.coverage.report]
show_missing = true
fail_under = 50

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_context = true
disable_error_code  = ["overload-overlap"]

[[tool.mypy.overrides]]
module = [
    "geopandas.*",
    "pyogrio.*",
    "fsspec.*",
    "requests.*",
]
ignore_missing_imports = true

[tool.ruff]
force-exclude = true  # Apply excludes to pre-commit
show-fixes = true
src = ["src", "tests"]
target-version = "py310"  # Minimum Python version supported
include = ["*.py", "*.pyi", "**/pyproject.toml", "*.ipynb"]
extend-exclude = [
    "__pycache__",
    "old",
    ".ipynb_checkpoints",
    "noxfile.py",
    "docs/conf.py",
]

# Ruff rules may be customized as desired: https://docs.astral.sh/ruff/rules/
[tool.ruff.lint]
select = [
    "A",    # prevent using keywords that clobber python builtins
    "ANN",  # check type annotations
    "B",    # bugbear: security warnings
    "D",    # documentation
    "E",    # pycodestyle
    "F",    # pyflakes
    "ISC",  # implicit string concatenation
    "I",    # sort imports
    "UP",   # alert you when better syntax is available in your python version
    "RUF",  # the ruff developer's own rules
]
ignore = [
    "A005",
    "ANN003", # Supress Missing type annotation for `**kwargs`
    "ANN401", # Allow type annotation with type Any.
    "D100",   # Supress undocumented-public-module. Only doc of public api required.
    "E402",   # Supress module-import-not-at-top-of-file, needed in jupyter notebooks.
    "E501",   # Supress line-too-long warnings: trust black's judgement on this one.
]

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.pydocstyle]
convention = "google"  # You can also use "numpy".

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["classmethod", "validator", "root_validator", "pydantic.validator"]

[tool.ruff.lint.per-file-ignores]
"*/__init__.py" = ["F401"]
"**/tests/*" = [
    "ANN001",  # type annotations don't add value for test functions
    "ANN002",  # type annotations don't add value for test functions
    "ANN003",  # type annotations don't add value for test functions
    "ANN201",  # type annotations don't add value for test functions
    "ANN204",  # type annotations don't add value for test functions
    "ANN205",  # type annotations don't add value for test functions
    "ANN206",  # type annotations don't add value for test functions
    "D100",    # docstrings are overkill for test functions
    "D101",
    "D102",
    "D103",
    "S101",    # asserts are encouraged in pytest
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
